### 2019-09-17 #1737 @csoneson
@alletsee, @dvanic, @ellessenne, @seabbs - would two of you be willing to review this submission for JOSS?
