### 2019-09-17 #1740 @whedon
[REVIEW]: ANCRTAdjust: An R package to adjust routine HIV testing data from antenatal care to reduce bias in estimating HIV prevalence trends
**Submitting author:** @brittanyblouin (<a href="http://orcid.org/0000-0001-7447-7467">Brittany Blouin</a>)
**Repository:** <a href="https://github.com/brittanyblouin/ANCRTAdjust" target ="_blank">https://github.com/brittanyblouin/ANCRTAdjust</a>
**Version:** v0.1.0
