### 2019-09-17 9d430ce
Added DOIs
