+++ paper/paper.bib
@@ -12,12 +12,12 @@
 	Author = {WHO},
 	Date-Added = {2019-07-23 19:58:57 +0000},
 	Date-Modified = {2019-07-23 20:00:30 +0000},
-	Institution = {Wordl Health Organization, Geneva (Switzerland)},
+	Institution = {World Health Organization, Geneva (Switzerland)},
 	Title = {Conducting HIV surveillance based on routine programme data among pregnant women attending antenatal clinics},
 	Year = {2015}}
 
 @article{Maheu-Giroux:2019,
-	Author = {Maheu-Giroux M., Jahn A., Kalua T., Mganga A., Eaton J.W.},
+  Author = {Maheu-Giroux M., Jahn A., Kalua T., Mganga A., Eaton J.W.},
 	Date-Added = {2019-07-23 19:44:33 +0000},
 	Date-Modified = {2019-07-23 19:46:56 +0000},
 	Journal = {AIDS},
@@ -26,6 +26,7 @@
 	Year = {2019}}
 
 @article{Diaz:2005,
+  doi = {10.1097/01.aids.0000172871.80723.3e},
 	Author = {Diaz, T. and De Cock, K. and Brown, T. and Ghys, P.D. and Boerma, J.T.},
 	Issn = {0269-9370},
 	Journal = {AIDS},
@@ -36,7 +37,7 @@
 	Year = {2005}}
 
 @techreport{WHO:2013,
-	Author = {WHO},
+  Author = {WHO},
 	Date-Modified = {2019-07-23 20:00:41 +0000},
 	Institution = {Wordl Health Organization, Geneva (Switzerland)},
 	Number = {9241505613},
