### 2019-09-17 #1740 @brittanyblouin
@csoneson - I have added the DOIs to the references.
