### 2019-09-26 #1740 @ellessenne
Hi @brittanyblouin (cc @csoneson),

First of all - well done with the package! I think you clearly describe the issue you aim to solve and the high-level functionality of ANCRTAdjust. The range of functions provided for end-users seems comprehensive (although I don't really work in the area, so hopefully I haven't missed anything obvious).
I have a few comments, some of which are quite specific and some more general ones.
You can find the former filed as issues in the ANCRTAdjust repository, and the more general comments follow here.

First, I think the documentation of the package could be improved: in particular, I think that the different arguments and options - albeit well described, strictly speaking - do not fully motivate their choice. What I am trying to say is, arguments are well described but they lack a bit of context and more methodological explanations: for instance, when using the `data_clean()` function it is not clear why one should choose an adjustment option over another. Another example is with the `mt_adjust()` function. I would suggest expanding the documentation a bit, and maybe including a vignette that describes the relevant methodology behind the functions implemented in this package.

Second, there are some tests in the 'test' folder - which is great! However, [`testthat`](https://testthat.r-lib.org) could be implemented to trigger automatic testing when running R CMD check.
More tests should be added to cover argument checks (and more argument checks should be implemented throughout the package as well).
The package *should* aim to pass R CMD check (I recommend running `devtools::check()` often) - do you aim to submit to CRAN eventually?

Third, there is no comparison with other software. Is there other software that can do something similar? If no, I would stress more the novelty of your work! If yes, a comparison needs to be included, focussing on strengths and weaknesses.

Fourth, I did not tick the "references" box yet: are the methods implemented in this package covered by the references included in the paper? Are they well established methods in the field? I feel I was not given enough information/details to assess this (and this somewhat connects with my first comment included above).

Overall, I like the package and the paper - the authors highlight the importance of standardised protocols and procedures to ensure consistency and reproducibility, and I share the same feeling.
However, I think the documentation of the package (and the paper, to a smaller extent) could be improved/expanded to describe methods in more detail and to be unambiguous to the end-user.

I hope this helps, and I look forward to discussing this in more detail!
Cheers,

Alessandro
