### 2019-09-30 brittanyblouin/ANCRTAdjust#13 @seabbs
Use of plyr 
It was my understanding that `plyr` had been retired (in favour of `dplyr`) with support really only there for legacy applications.

Could you give me some justification for using it? Alternatively switching out to `dplyr` would be great.

For this #openjournals/joss-reviews/issues/1740 review
