### 2019-09-30 brittanyblouin/ANCRTAdjust#14 @seabbs
JOSS paper
Overall this is a good paper but I think it needs a few tweaks to make the messaging clearer.

## Statement of need

Data cleaning is very important and missing data can be a major issue with surveillance data. I think these points come across well in your paper

I am a little concerned as to the utility of this package given it can only work with a single data format. How widespread is this data format? Is this a toolbox that many users can easily pick up and use? Why not make a more general data cleaning package that could be used with this dataset as an example?

What I am getting at here is what is the special features of this dataset that means that it needs its own data cleaning package rather than off the shelf tooling?

You discuss the fact that having clear and simple data handling guidelines is a good idea in the paper but the package doesn't really document this process at the moment. Instead it looks to me like a series of tools that someone could use to make such a guide.

I think it would make the paper and documentation stronger if you dealt with these points.

## State of the field

Agree with @ellessenne here. Really need to see some comparison to other tools.

This needs to be in two forms:

 1.) What tools exist already for high level data cleaning and why is this standard tool kit required for this data.

2.) Specific package functionality - why is this your imputation of missing data for example better than other commonly used methods (i.e MICE).

This needs to be in the package documentation as well.

## Language issues

There are few language issues in the paper that make understanding it difficult - language could use some tightening.

### Example

 *"If ANC-RT data on HIV serostatus are to be used for HIV surveillance, several challenges need to be addressed, however (Diaz, De Cock, Brown, Ghys, & Boerma, 2005). First, inferences from routinely collected program data could be biased due to imperfect data completeness, with some health facilities not reporting HIV testing data (WHO, 2013). "*

For this #[openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
