### 2019-09-30 brittanyblouin/ANCRTAdjust#15 @seabbs
Checking data structure
As far as  I can see this package is very reliant on having a very specific data structure. Given this I think it is very important that you check the input data in some way before trying to use it.

This could take the form of checks within your already implemented functions or a separate function to check the data integrity.

For this # [openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
