### 2019-09-30 brittanyblouin/ANCRTAdjust#16 @seabbs
Using the data cleaning tools - vignette
As discussed in the paper this package is meant to provide guidelines for clear and simple data handling.

To do this more fully a walkthrough is needed starting with a dataset and stepping through each step.  This would give an analyst with some data a pipeline to follow rather than the current situation where you are giving them a toolbox of tools to use.

Having documentation in this pipeline giving a high level overview of what each function does would also be great (this high level explanation is also needed in the function documentation itself).

An example of this is `data_clean` which does quite a few things - not all of which are immediately obvious from a quick scan of the present docs. Having this set up also means I need to go from docs to function docs and back just to get an understanding of what the package is doing - which is not ideal.

Linked to this #14 issue.

For this # [openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
