### 2019-09-30 brittanyblouin/ANCRTAdjust#17 @seabbs
Testing
As discussed in the draft paper this package is aiming to be a standardised toolbox for cleaning a specific data format. To give others users confidence all functionality needs to be thoroughly tested (both at a high level with examples and at a low unit test level).

Whilst there is currently a `test.R` script as it currently stands it isn't really testing anything (apart from checking for errors). To act as  a tests expected output needs to be given (or interpretation of plots).

I would like to see unit tests using `testthat` and `vdiffr` for all low level functionality. I would also like to  see a high level walkthrough going through what is actually happening on an example dataset.

[Here](https://github.com/seabbs/getTBinR/tree/master/tests/testthat) is an example of some unit tests I have previously implemented to get you started. Happy to give more details if needed.

Linked to #16

As mentioned by @ellessene in their review.

For this # [openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
