### 2019-09-30 brittanyblouin/ANCRTAdjust#18 @seabbs
Functionality
I need more evidence before I can say the package works as intended. This should be in the form of tests, documentation and an example pipeline (all mentioned in other issues).

Linked to #17 and #16

For this # [openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
