### 2019-09-30 brittanyblouin/ANCRTAdjust#19 @seabbs
descriptive_plot axis limits
Do these need to be manually specified?

The current default shows the complete percentage range which in most circumstances is not going to be ideal when trying to understand trends over time (which is my interpretation of what this function is doing).

My preference would be to have no specified range and rely on `ggplot2` defaults. User can then add limits as required.

What are your thoughts on this?

For this # [openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
