### 2019-09-30 brittanyblouin/ANCRTAdjust#21 @seabbs
CRAN release
This is not a blocker for JOSS publication but are there any plans for a CRAN release of this software?

In general being on CRAN is an important marker of quality (at least basic quality) and makes installation much easier.

For this # [openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
