### 2019-09-30 brittanyblouin/ANCRTAdjust#15 @seabbs
After rereading the README I see some of this functionality is in place in the `name_var` function.

I think you should add these checks (via calling this function) to all other functions (or at least the `data_clean`).

A name change might help as well - perhaps something like `check_data_structure`.

I think you also need to be checking that the variables supplied have the required structure (on top of just names). By this I mean are they the correct type (numeric etc) and do they have the right levels (for factors).

For this #[openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
