### 2019-09-30 #1740 @seabbs
Hi @brittanyblouin (cc @csoneson) ,

Well done @brittanyblouin this is a nice package - we definitely need to see more of these kinds of tools for standardising data handling.

Before I can recommend this package though I think there needs to be some more work tightening up - and expanding - the documentation. The rigour of the coding (i.e testing, function names, environment variable handling etc.) also needs some work.

Looking forward to getting into  the details!

Very much seconding all of @ellessenne points in my review as their review was already pretty watertight (nice job 👍). Expanded on them in the review issues where I felt I had something to add.
