### 2019-09-30 brittanyblouin/ANCRTAdjust#24 @seabbs
Spell check
Found a few spelling issues in the docs - although far fewer that in mine!

I recommend the spelling package for automatic spell checking. See [here](https://github.com/seabbs/getTBinR/blob/master/tests/spelling.R) for an example implementation.

For this # [openjournals/joss-reviews/issues/1740](https://github.com/openjournals/joss-reviews/issues/1740)  review
