### 2019-10-05 #1740 @csoneson
@ellessenne, @seabbs - thanks a lot for your comments!
@brittanyblouin - please go through the comments above and the issues that have been opened in your repository - don't hesitate to ping me and/or continue the discussion in the respective issue if you have any questions, and please report back here when you are ready for the reviewers to take another look!
