### 2019-10-07 #1740 @brittanyblouin
Hi @ellessenne and @seabbs, Thank you so much for the comments and review of our package.  We will go over all comments and issues shortly and get back with an improved version.
