### 2019-10-07 #1740 @brittanyblouin
Hi @csoneson - Can we add co-author @m-maheu-giroux as a corresponding author for this submission?
