### 2019-10-07 #1740 @csoneson
> Hi @csoneson - Can we add co-author @m-maheu-giroux as a corresponding author for this submission?

@brittanyblouin Since @m-maheu-giroux is mentioned in this issue now, I believe he will be notified of all further activity here (which would also be achieved by clicking the "Subscribe" button above). Would that be a satisfactory solution? Or do you rather mean that you would like to hand over the "submitting author" status?
