### 2019-10-07 #1740 @m-maheu-giroux
Yes, this is a satisfactory solution (thanks!).
Brittany Blouin shall remain "submitting author".
