+++ README.md
@@ -37,7 +37,7 @@ data(ancrt)
 
 ``` r
 > head(ancrt)
-  faciluid time n_clients n_status knownpos testpos testneg  true_prv snu1 Year
+  faciluid time n_clients n_status knownpos testpos testneg  true_prv snu1 year
 1      F_1    1       217      217       29      37     151 0.3041475    1 2015
 2      F_1    2       240      240       31      40     169 0.2958333    1 2015
 3      F_1    3       216      216       30      38     148 0.3148148    1 2015
@@ -66,12 +66,12 @@ Some data might be available age-disaggregated. If so, the variable should be na
 Finally, data might be available for the sub-national unit 1 and for the year (allowing some results to be stratified by these variables).  If so, these variables should be named:
 
 * `snu1`: The sub-national unit 1.
-* `Year`: The calendar year that the data was collected.
+* `year`: The calendar year that the data was collected.
 
 The function `name_var()` can automatically rename the variables for you (see `help(name_var)` for more information).
 
 ``` r
-> ancrt <- name_var(ancrt, faciluid = "faciluid", time = "time", n_clients = 'n_clients', n_status = "n_status", knownpos = "knownpos", testpos = "testpos", testneg = "testneg", snu1 = "snu1", Year = "Year")
+> ancrt <- name_var(ancrt, faciluid = "faciluid", time = "time", n_clients = 'n_clients', n_status = "n_status", knownpos = "knownpos", testpos = "testpos", testneg = "testneg", snu1 = "snu1", year = "year")
 ```
 
 Data cleaning
@@ -140,7 +140,7 @@ ancrt_cleaned <- mt_adjust(ancrt_cleaned, adjust_option = 'setmax')
 
 Flagging outliers
 -----------------
-It is possible that errors exist in the data and these may be detected by investigating outlier observations.  The function `flag_outliers()` will flag outliers for the variables `n_clients`, `n_status_c`, `testpos_c`, `testneg_c`, `knownpos_c`, `totpos_c`, `Prv` and `Cov`.  Outliers are defined as being more than two standard deviations greater than or less than the mean value.  The mean reference to which observations are compared can either be the facility-specific mean, the subnational unit 1-specific mean or the country-specific mean, according to user inputs.  The flagged data can either be output as the full dataset with additional variables indicating outlier observations, or a dataset that only includes data for outlier observations.  See `help(flag_outliers)` for more information.
+It is possible that errors exist in the data and these may be detected by investigating outlier observations.  The function `flag_outliers()` will flag outliers for the variables `n_clients`, `n_status_c`, `testpos_c`, `testneg_c`, `knownpos_c`, `totpos_c`, `prv` and `cov`.  Outliers are defined as being more than two standard deviations greater than or less than the mean value.  The mean reference to which observations are compared can either be the facility-specific mean, the subnational unit 1-specific mean or the country-specific mean, according to user inputs.  The flagged data can either be output as the full dataset with additional variables indicating outlier observations, or a dataset that only includes data for outlier observations.  See `help(flag_outliers)` for more information.
 
 ``` r
 outliers <- flag_outliers(ancrt_cleaned, flagby = "facility", result = "outliers")
@@ -148,7 +148,7 @@ outliers <- flag_outliers(ancrt_cleaned, flagby = "facility", result = "outliers
 
 ``` r
 > head(outliers)
-  faciluid time n_clients n_status_c testpos_c testneg_c knownpos_c totpos_c       Prv       Cov
[…]
