### 2019-10-22 bbe4855
Integrating reviewers' comments
Naming convention standardized
impcov_adjust() funtions combined
Adding tests to testthat
Updated readme
