### 2019-10-23 brittanyblouin/ANCRTAdjust#13 @m-maheu-giroux
Our understanding was that `dplyr` cannot accommodate user defined functions at this time. Hence, we went with `plyr`.
