### 2019-10-23 brittanyblouin/ANCRTAdjust#19 @m-maheu-giroux
We also initially wondered if it would be best to just use the default but decided against this. The issue being that prevalence and testing coverage could potentially exhibit important variations over time and across facilities. Comparing them using standardized default values makes more sense.
