+++ tests/testthat/test_function.R
@@ -1,18 +1,20 @@
 context("test_plot_function.R")
 
-data("ancrt")
+ANCRTAdjust::ancrt
 
-ancrt_cleaned <- data_clean(ancrt)
+ancrt_cleaned <- ANCRTAdjust::data_clean(ancrt)
 
-test_that("descriptive_plot produces a plot", {
-  plot <- descriptive_plot(ancrt_cleaned)
+testthat::test_that("descriptive_plot produces a plot", {
+  plot <- ANCRTAdjust::descriptive_plot(ancrt_cleaned)
   expect_true(!is.null(plot))
   expect_equal("ggplot", class(plot)[2])
 })
 
-prev_cov <- HIVprev_ipcw(ancrt_cleaned)
-results <- impcov_adjust(prev_cov)
+prv_cov <- ANCRTAdjust::hiv_prv_ipcw(ancrt_cleaned, by_snu1 = FALSE, by_period = TRUE, by_year = FALSE)
+results <- ANCRTAdjust::impcov_adjust(data = prv_cov)
 
-test_that("descriptive_plot produces a plot", {
-  plot <- plot_rawadjusted(results, snu1 = "All", time.unit = "PERIOD", HIVraw = "TRUE", y.lim = 40)
+testthat::test_that("plot_rawadjusted produces a plot", {
+  plot <- ANCRTAdjust::plot_rawadjusted(data = results, snu1 = "all", time_unit = "period", hiv_raw = TRUE, y_lim = 40)
+  expect_true(!is.null(plot))
+  expect_equal("ggplot", class(plot)[2])
 })
