+++ man/ancrt.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ancrt.R
+\docType{data}
+\name{ancrt}
+\alias{ancrt}
+\title{Test ANC-RT dataset.}
+\format{A data frame with 20,000 rows and 10 variables:
+\describe{
+  \item{faciluid}{the unique facility identifier}
+  \item{time}{the time period}
+  \item{n_clients}{the number of clients attending antenatal care (ANC)}
+  \item{n_status}{the number of women whose HIV status was ascertained}
+  \item{knownpos}{the number of women who were already diangosed as living with HIV}
+  \item{testpos}{the number of women who tested positive to HIV at ANC}
+  \item{testneg}{the number of women who tested negative to HIV at ANC}
+  \item{true_prv}{the true underlying HIV prevalence at this facility for the simulated dataset}
+  \item{snu1}{the identifier of the first sub-national unit}
+  \item{year}{the calendar year of the observation}
+}}
+\usage{
+ancrt
+}
+\description{
+A simulated dataset containing information on the number of clients attending antenatal care (ANC), the number for which their HIV status was ascertained, the number of known positive, the number who tested positive, the number who tested negative for eahc facility and time period.
+}
+\keyword{datasets}
