+++ README.Rmd
@@ -1,3 +1,7 @@
+---
+output: html_document
+---
+
 ANCRTAdjust
 =======
 
