+++ R/hiv_prv_cov.R
@@ -6,12 +6,12 @@
 #' The HIV prevalence and testing coverage can be reported stratified by reporting period, year and/or subnational unit 1, 
 #' according to user inputs.
 #'
-#' @param data The ANC-RT dataset.  The functions \link[ANCRTAdjust]{name_var()}, \link[ANCRTAdjust]{data_clean()} and \link[ANCRTAdjust]{mt_adjust()} should have been run on the data to properly
-#' prepare the data for use here.  The dataset must have the following variables:
+#' @param data The ANC-RT dataset. The functions \link[ANCRTAdjust]{name_var}, \link[ANCRTAdjust]{data_clean} and \link[ANCRTAdjust]{mt_adjust} should have been run on the data to properly
+#' prepare the data for use here. The dataset must have the following variables:
 #'  \itemize{
 #'   \item \code{n_clients}: The number of women from the specified facility, during the specified time period, that attended their first ANC visit.
-#'   \item \code{totpos_c}: Cleaned \code{totpos} (generated using the \link[ANCRTAdjust]{data_clean()} and \link[ANCRTAdjust]{mt_adjust()} functions).
-#'   \item \code{n_status_c}: Cleaned \code{n_status} (generated using the \link[ANCRTAdjust]{data_clean()} and \link[ANCRTAdjust]{mt_adjust()} functions).
+#'   \item \code{totpos_c}: Cleaned \code{totpos} (generated using the \link[ANCRTAdjust]{data_clean} and \link[ANCRTAdjust]{mt_adjust} functions).
+#'   \item \code{n_status_c}: Cleaned \code{n_status} (generated using the \link[ANCRTAdjust]{data_clean} and \link[ANCRTAdjust]{mt_adjust} functions).
 #'   \item \code{snu1}: The subnational unit 1 (only required if results are to be stratified by snu1).
 #'   \item \code{time}: The time period over which the data was collected (only required if results are to be stratified by the reporting time period).
 #'   \item \code{year}: The year that the data was collected (only required if results are to be stratified by year).
@@ -37,7 +37,7 @@
 #' 
 
 hiv_prv_cov <- function(data, by_period = FALSE, by_snu1 = FALSE, by_year = FALSE) {
-
+    
   hiv_prvs <- function(data) {
     raw <- round(((weighted.mean((data$totpos) / data$n_status, w = data$n_status, na.rm = TRUE)) * 100), 2)
     prv <- round(((weighted.mean((data$totpos_c) / data$n_status_c, w = data$n_status_c, na.rm = TRUE)) * 100), 2)
