+++ R/plot_rawadjusted.R
@@ -10,15 +10,15 @@
 #'  \itemize{
 #'   \item \code{snu1}: The subnational unit 1.  If plots are to be generated for the different snu1s, the parameter by_snu1 = "TRUE" must 
 #'   have been specified when running \link[ANCRTAdjust]{impcov_adjust}.
-#'   \item \code{time}: The year and/or quarter that the data was collected.If plots are to be generated by reporting period, the parameter 
-#'   by_period = "TRUE" must have been specified when running \link[ANCRTAdjust]{impcov_adjust}. If plots are to be generated by year, the parameter
-#'   by_year = "TRUE" must have been specified when running \link[ANCRTAdjust]{impcov_adjust}.
+#'   \item \code{time}: The year and/or period (i.e., quarters) that the data was collected.If plots are to be generated by reporting period, the parameter 
+#'   by_period = TRUE must have been specified when running \link[ANCRTAdjust]{impcov_adjust}. If plots are to be generated by year, the parameter
+#'   by_year = TRUE must have been specified when running \link[ANCRTAdjust]{impcov_adjust}.
 #'   \item \code{adjusted_prv}: The HIV prevalence adjusted for imperfect testing coverage and all previous adjustments (i.e. data cleaning, adjustment 
 #'   for multiple testing and/or adjustment for missing reporting periods, if performed) (generated using the \link[ANCRTAdjust]{impcov_adjust} function).
 #'   \item \code{hiv_raw}: The HIV prevalence using the raw data (generated using the \link[ANCRTAdjust]{impcov_adjust} function).
 #'  }
-#' @param snu1 The snu1 for which the plot is generated. "All" indicates that the plot include the entire country data (i.e. not snu1-specific). Default = "All".
-#' @param time_unit "PERIOD" indicates that the results be plotted by reporting period.  "YEAR" indicates that the results be plotted by year. Default = "PERIOD".
+#' @param snu1 The snu1 for which the plot is generated. "all" indicates that the plot include the entire country data (i.e. not snu1-specific). Default = "all".
+#' @param time_unit "period" indicates that the results be plotted by reporting period. "year" indicates that the results be plotted by year. Default = "period".
 #' @param hiv_raw "TRUE" indicates that both the fully adjusted and raw HIV prevalences be plotted.  "FALSE" indicates that only the fully
 #' adjusted HIV prevalences be plotted. Default = "TRUE".
 #' @param y_lim The y-axis upper limit. Default = 40.
@@ -32,9 +32,16 @@
 #'
 #' @export
 
-plot_rawadjusted <- function(data, snu1 = "All", time_unit = "PERIOD", hiv_raw = "TRUE", y_lim = 40){
+plot_rawadjusted <- function(data, snu1 = "all", time_unit = "period", hiv_raw = TRUE, y_lim = 40) {
+  if (!(time_unit %in% c('period', 'year'))) {
+    stop("time_unit must be eithe 'period' or 'year'")
+  }
+  
+  adjusted_prv <- NULL
+  
   data$time <- as.numeric(data$time)
-  if(time_unit == "PERIOD" & hiv_raw == "TRUE"){
+  
+  if(time_unit == "period" & hiv_raw == TRUE) {
     Plot <-
     ggplot(data[data$time < 9000 & data$snu1 == snu1,], aes(time)) +
     geom_line(aes(y = hiv_raw, color = "Raw HIV prevalence"), size=1) + geom_point(aes(y = hiv_raw, color = "Raw HIV prevalence")) +
[…]
