+++ README.md
@@ -203,7 +203,7 @@ Step 8) Adjusting HIV prevalence for imperfect testing coverage
 -------------------------------------------------------
 It has been found that selection bias can be introduced into HIV prevalence estimates due to imperfect HIV testing coverage (i.e. < 100% coverage). If the average testing coverage is low (i.e., <90%), it can be advised to adjust for these potential biases.
 
-Using data from the *President Emergency Funds for AIDS Relief in Africa*'s (PEPFAR) *Monitoring, Evaluation, and Reporting* (MER) database, we estimated adjustment factors for each 1% point increase in HIV testing coverage. To do so, we used binomial logistic regression models with facility-level fixed effects and marginal standardization to assess the effect of testing coverage on HIV prevalence (see Maheu-Giroux et al. (2019) for details on the methods, applied to Malawi's ANC-RT data). The PEPFAR MER database contains information on more than 37 millions ANC attendees from 19,527 unique facilities from 17 countries in sub-Saharan Africa, totalling 226,541 observations over the 2015-2019 period. The figure below shows the estimated
+Using data from the *President Emergency Funds for AIDS Relief in Africa*'s (PEPFAR) *Monitoring, Evaluation, and Reporting* (MER) database, we estimated adjustment factors for each 1% point increase in HIV testing coverage. To do so, we used binomial logistic regression models with facility-level fixed effects and marginal standardization to assess the effect of testing coverage on HIV prevalence (see Maheu-Giroux et al. (2019) for details on the methods, applied to Malawi's ANC-RT data). The PEPFAR MER database contains information on more than 37 millions ANC attendees from 19,527 unique facilities from 17 countries in sub-Saharan Africa, totalling 226,541 observations over the 2015-2019 period. The figure below shows the estimated relationship between imperfect testing coverage and HIV preavlence (Fig 2)
 
 ![Fig. 2. Relationship between routine HIV testing coverage during antenatal care and predicted HIV prevalence. The dots on the graph are a subset of monthly observations of HIV prevalence at antenatal care. The size of the dots is proportional to the number of women attending antenatal care.](man/figures/PEPFAR_MER.png)
 
