### 2019-10-24 brittanyblouin/ANCRTAdjust#13 @seabbs
The `mutate` , `purrr::map` and then `unnest supports this. Happy to walk through in more detail if needed or explore other options.
