### 2019-10-24 brittanyblouin/ANCRTAdjust#19 @seabbs
Okay - happy to go with you on this.

However, it can be difficult to see both trends over time and differences between groups if the groups are on very difficult scales. The current plot obscures trends if there are large differences. Might want to discuss this a little in the docs or provide a standardised option (for trends vs. differences).
