### 2019-10-24 brittanyblouin/ANCRTAdjust#19 @m-maheu-giroux
Fair point. I have implemented your requested changes and switched to default. Users still have the option of using their own supplied limits for the y-axis
