+++ R/descriptive_plot.R
@@ -32,9 +32,9 @@
 #'   \item \code{totpos_c_setmax}: Adjusted \code{totpos_c} if the set to maximum adjustment option for multiple testing is used (generated using the \link[ANCRTAdjust]{data_clean} function)
 #'   \item \code{time}: The time period
 #'  }
-#' @param ylim_ind The y-axis lower and upper limits of the plot of the primary data quality indicators over time (default = c(0, 100))
-#' @param ylim_cov The y-axis lower and upper limits of the plot of HIV testing coverage over time (default = c(0, 100))
-#' @param ylim_prv The y-axis lower and upper limits of the plot of HIV prevalence over time (default = c(0, 100))
+#' @param ylim_ind The y-axis lower and upper limits of the plot of the primary data quality indicators over time (if left blank, ggplot() defaults are used)
+#' @param ylim_cov The y-axis lower and upper limits of the plot of HIV testing coverage over time (if left blank, ggplot() defaults are used)
+#' @param ylim_prv The y-axis lower and upper limits of the plot of HIV prevalence over time (if left blank, ggplot() defaults are used)
 #' @param plot_type The plots included in the final plot.  Options include:
 #'  \itemize{
 #'   \item \code{full} All three plots are included (default)
@@ -57,10 +57,11 @@
 #'
 #' @export
 
-descriptive_plot <- function (data, 
-                              ylim_ind = c(0, 100), 
-                              ylim_cov = c(0, 100), 
-                              ylim_prv = c(0, 100), plot_type = "full") {
+descriptive_plot <- function (data, plot_type = "full") {
+  
+  if (is.null(data$n_status_c)) {
+    stop("Error: please use the data_clean() function prior to producing descriptive plots")
+  }
 
   data$totpos_raw <- ifelse(!is.na(data$knownpos) & !is.na(data$testpos), data$knownpos + data$testpos, 
                             data$totpos)
@@ -143,11 +144,12 @@ descriptive_plot <- function (data,
     geom_point(aes(y = missingdata_cleaned, color = "Missing data in cleaned data")) +
     xlab("") +
     ylab("% facilities with data \n quality problem") +
-    ylim(ylim_ind) +
     theme(axis.title.y = element_text(size = 9)) +
     scale_colour_manual(name = "",
                         values = c("Invalid values in raw data" = "orange", "Missing data in cleaned data" = "lightcoral")) +
     geom_abline(aes(intercept = 10, slope = 0), linetype = "dashed")
+  if (exists("ylim_ind")) { 
+    indicator_plot <- indicator_plot + ylim(ylim_ind) }
   
   time <- coverage <- coverage_impute <- coverage_setmax <- coverage_remove <- coverage_raw <- NULL
   coverage_plot <-
@@ -162,7 +164,6 @@ descriptive_plot <- function (data,
     geom_point(aes(y = coverage_remove, color = "Remove option")) +
     geom_line(aes(y = coverage_raw, color = "Raw data"), size = 1) + 
     geom_point(aes(y = coverage_raw, color = "Raw data")) +
-    ylim(ylim_cov) +
     xlab("") +
     ylab("HIV testing coverage \n (%)") +
     theme(axis.title.y = element_text(size = 9)) +
@@ -170,7 +171,9 @@ descriptive_plot <- function (data,
                       values = c("Raw data" = "red", "Cleaned data" = "darkgoldenrod4",
                                  "Set to maximum option         " = "purple", "Remove option" = "blue", "Impute option" = "forestgreen")) +
[…]
