### 2019-10-24 brittanyblouin/ANCRTAdjust#18 @m-maheu-giroux
We have extensively tested the functions on ANC-RT data from the *President Emergency Funds for AIDS Relief in Africa*'s (PEPFAR) *Monitoring, Evaluation, and Reporting* (MER) database that contains information on more than contains information on more than 37 millions ANC attendees from 19,527 unique facilities from 17 countries in sub-Saharan Africa, totaling 226,541 observations over the 2015-2019 period. We also informally tested the package/functions on ANC-RT data from a few selected countries.

Unfortunately this data is not in the public domain. However, we summarized our findings in a 63 pages report that we can privately share with you. Ideally, we would be able to make this report public in the near future.
