### 2019-10-24 brittanyblouin/ANCRTAdjust#17 @m-maheu-giroux
Thank you for this suggestion. You are correct to have pointed out this omission in our first submission. We have implemented the main tests and everything is running smoothly.
