### 2019-10-24 brittanyblouin/ANCRTAdjust#21 @m-maheu-giroux
In the medium term, yes, we intend to submit to CRAN. However, we would first like to obtain users feedback on functions before doing so. Publishing the package on JOSS and github first seems like the appropriate sequence.
