### 2019-10-24 brittanyblouin/ANCRTAdjust#16 @m-maheu-giroux
We agree with your assessment and have considerably re-written the README to reflect your concerns. We now provide detailed step-by-step information on the data cleaning and analysis pipeline.
