### 2019-10-24 brittanyblouin/ANCRTAdjust#15 @m-maheu-giroux
Good idea.
We have implemented additional check to ensure that the data is correctly formatted and, to reflect this change, have renamed the function `name_var()` to `check_data()`.
