### 2019-10-24 brittanyblouin/ANCRTAdjust#14 @m-maheu-giroux
You raise valid concerns. Here is how I have addressed them.

## Statement of need
We have built our functions around the most commonly encountered data format. We have requested and trialed our packages on several ANC-RT datasets from sub-Saharan African countries. This format for the dataset is very common as the great majority of countries on this continent are using this information to assess their HIV epidemics trends, updating each year the UNAIDS-supported Spectrum/EPP software used for monitoring HIV epidemics.

This is also the very reason why cleaning and adjusting this type of data requires is own software: we need to specifically address the idiosyncrasies of the ANC-RT data and their potential biases. We have attempted to better reflect the importance of this tool in both the JOSS paper and the package information.

## State of the field
We have modified the paper to better reflect the novelty of this tool.

## Language issues
We simplified this section of the paper to improve understanding.
