+++ R/flag_outliers.R
@@ -5,7 +5,7 @@
 #' This function has been developed to flag outlier observations for the following variables: \code{n_clients}, \code{n_status_c}, \code{testpos_c}, \code{testneg_c}, \code{knownpos_c}, 
 #' \code{totpos_c}, \code{prv} and \code{cov}.  Outliers are defined as 2 standard deviations greater than or less than the mean value.  
 #' 
-#' @param data The ANC-RT dataset.  The functions \link[ANCRTAdjust]{name_var}, \link[ANCRTAdjust]{data_clean} and \link[ANCRTAdjust]{mt_adjust} should have been run on the data to properly
+#' @param data The ANC-RT dataset.  The functions \link[ANCRTAdjust]{check_data}, \link[ANCRTAdjust]{data_clean} and \link[ANCRTAdjust]{mt_adjust} should have been run on the data to properly
 #' prepare the data for use here.  The dataset must have the following variables:
 #'  \itemize{
 #'   \item \code{faciluid}: Facility ID.
@@ -24,6 +24,7 @@
 #'   \item \code{cov}:  The HIV testing coverage from the specified facility at the specified time period (generated using the \link[ANCRTAdjust]{mt_adjust} function).
 #'   \item \code{snu1}: The subnational unit 1 (only required if results are to be flagged by snu1).
 #'    }
+#' @param sd_for_outlier Standard deviation used to flag outliers (default is 2).
 #' @param flag_by Options include:
 #'  \itemize{
 #'    \item "\code{facility}" compares each observation's value to their facility's mean value and flags the observations that 
@@ -59,7 +60,7 @@
 #' @export
 #' 
 
-flag_outliers <- function(data, flag_by = "facility", result = "outliers") {
+flag_outliers <- function(data, sd_for_outlier = 2, flag_by = "facility", result = "outliers") {
   
   flag_n_clients <- faciluid <- n_clients <- flag_n_status_c <- n_status_c <- flag_testpos_c <- NULL
   testpos_c <- flag_testneg_c <- testneg_c <- flag_knownpos_c <- knownpos_c <- flag_totpos_c <-NULL
@@ -69,22 +70,22 @@ flag_outliers <- function(data, flag_by = "facility", result = "outliers") {
     data1 <- NULL
     for (i in levels(as.factor(data$faciluid))) {
       temp <- data[data$faciluid == i,]
-      temp$flag_n_clients <- ifelse(((temp$n_clients > (mean(temp$n_clients, na.rm = TRUE) + (2 * sd(temp$n_clients, na.rm = TRUE)))) & !is.na(temp$n_clients)) | 
-                                    ((temp$n_clients < (mean(temp$n_clients, na.rm = TRUE) - (2 * sd(temp$n_clients, na.rm = TRUE)))) & !is.na(temp$n_clients)), 1, 0)
-      temp$flag_n_status_c <- ifelse(((temp$n_status_c > (mean(temp$n_status_c, na.rm = TRUE) + (2 * sd(temp$n_status_c, na.rm = TRUE)))) & !is.na(temp$n_status_c)) | 
[…]
