+++ R/quality_indicators.R
@@ -148,7 +148,7 @@ quality <- function(data) {
 #' \code{testneg} and, \code{knownpos}) are calculated and output in a dataframe. Data quality indicators are calculated from both raw and cleaned
 #' variables and can be calculated stratified by region and time, according to user inputs.
 #' 
-#' @param data The ANC-RT dataset.  The functions \link[ANCRTAdjust]{name_var} and \link[ANCRTAdjust]{data_clean} should have been run on the data to properly
+#' @param data The ANC-RT dataset.  The functions \link[ANCRTAdjust]{check_data} and \link[ANCRTAdjust]{data_clean} should have been run on the data to properly
 #' prepare the data for use here.  The data set must have the following variables:
 #'  \itemize{
 #'   \item \code{faciluid}: The facility identification code
@@ -212,14 +212,26 @@ quality_indicators <- function(data, by_region = FALSE, by_time = FALSE){
     
     table2 <- NULL
     region.list <- (unique(data$snu1))
-    for (i in 1:length(region.list)) {
-      RegionData <- data[data$snu1 == region.list[i],]
+    for (i in region.list) {
+      RegionData <- data[data$snu1 == i, ]
       table3 <- quality(RegionData)
       table3$region <- region.list[i]
-      row.names(table3)<-c(paste("Missing >=1 quarter", i, sep = '.'), paste("Missing n_clients", i, sep = '.'), paste("Missing n_status", i, sep = '.'), paste("Missing testpos", i, sep = '.'), paste("Missing testneg", i, sep = '.'),
-                           paste("Missing knownpos", i, sep = '.'), paste("Missing >=1 variables", i, sep = '.'), paste("Invalid coverage", i, sep = '.'), paste("Invalid prevalence", i, sep = '.'),
-                           paste("Inconsistent n_status", i, sep = '.'), paste("Negative n_clients", i, sep = '.'), paste("Negative n_status", i, sep = '.'), paste("Negative testpos", i, sep = '.'), paste("Negative testneg", i, sep = '.'),
-                           paste("Negative knownpos", i, sep = '.'), paste("One or more invalid variables", i, sep = '.'))
+      row.names(table3) <- c(paste("Missing >=1 quarter", " (region ", i, ")", sep = "" ), 
+                             paste("Missing n_clients", " (region ", i, ")", sep = "" ), 
+                             paste("Missing n_status", " (region ", i, ")", sep = "" ), 
+                             paste("Missing testpos", " (region ", i, ")", sep = "" ), 
+                             paste("Missing testneg", " (region ", i, ")", sep = "" ),
+                           paste("Missing knownpos", " (region ", i, ")", sep = "" ), 
+                           paste("Missing >=1 variables", " (region ", i, ")", sep = "" ), 
+                           paste("Invalid coverage", " (region ", i, ")", sep = "" ), 
+                           paste("Invalid prevalence", " (region ", i, ")", sep = "" ),
+                           paste("Inconsistent n_status", " (region ", i, ")", sep = "" ), 
+                           paste("Negative n_clients", " (region ", i, ")", sep = "" ), 
+                           paste("Negative n_status", " (region ", i, ")", sep = "" ), 
+                           paste("Negative testpos", " (region ", i, ")", sep = "" ), 
+                           paste("Negative testneg", " (region ", i, ")", sep = "" ),
+                           paste("Negative knownpos", " (region ", i, ")", sep = "" ), 
+                           paste("One or more invalid variables", " (region ", i, ")", sep = "" ))
       table2 <- rbind(table2, table3)
     }
     table <- rbind(table, table2)
[…]
