+++ man/quality_indicators.Rd
@@ -7,7 +7,7 @@
 quality_indicators(data, by_region = FALSE, by_time = FALSE)
 }
 \arguments{
-\item{data}{The ANC-RT dataset.  The functions \link[ANCRTAdjust]{name_var} and \link[ANCRTAdjust]{data_clean} should have been run on the data to properly
+\item{data}{The ANC-RT dataset.  The functions \link[ANCRTAdjust]{check_data} and \link[ANCRTAdjust]{data_clean} should have been run on the data to properly
 prepare the data for use here.  The data set must have the following variables:
  \itemize{
   \item \code{faciluid}: The facility identification code
