### 2019-10-24 9dfca7c
small update to documentation
