### 2019-10-24 #1740 @m-maheu-giroux
Hi @ellessenne and @seabbs,

First, thanks a lot for your thorough review. We have yet to publish many R packages and we really appreciate the time and efforts it took for you to revise our work (and to correct some of the omissions/blunder we made).

We believe that we have addressed all your concerns and suggestions. I have not closed the issues you raised yet as I feel it is your prerogative to do so.

Do not hesitate to contact me if you have any further questions. Thanks again.

Best,
Mathieu MG
