### 2019-10-24 brittanyblouin/ANCRTAdjust#13 @m-maheu-giroux
I initially closed them because it was easier for me to go through your comments... but the point is to facilitate your work - not mine. I have left the other issues for you to close.
