### 2019-10-25 brittanyblouin/ANCRTAdjust#13 @seabbs
No problem at all - thanks for doing that.

Did you have any thoughts on switching to `dplyr`. I had another look at your code base and it does all look pretty do-able.

@ellessenne do you agree with me on this or is the current implementation okay?
