### 2019-10-25 brittanyblouin/ANCRTAdjust#24 @seabbs
Great - I assume that the spelling issues both reviewers identified in pull requests have been sorted?
