### 2019-10-25 brittanyblouin/ANCRTAdjust#18 @seabbs
JOSS has a policy that all confirmatory evidence needs to be in the public domain and be open access.

I can totally understand that there may be data access issues. Is synthetic data a viable alternative? Ultimately your users need to have evidence that the package works as intended.
