### 2019-10-25 brittanyblouin/ANCRTAdjust#21 @seabbs
Great - good to hear.

Have you checked that everything is passing CRAN checks? (I use `devtools::check_rhub()` for this)
