### 2019-10-25 brittanyblouin/ANCRTAdjust#16 @seabbs
This is much improved. New layout gives me a much better idea of the pipeline.
