### 2019-10-25 brittanyblouin/ANCRTAdjust#18 @seabbs
Looking again at the new README (which is much improved) has made the package functionality much clearer!

I just need a few more things:

* Evidence the synthetic data your are using has the same problems you would see in raw data. This could just be details on its generation.
* Fully automated testing with full coverage.
* A discussion on why mean imputation has been used. Mean imputation is relatively crude in comparison to multiple imputation or other modelling approaches so I think some justification is needed.

Nice to have:

* An evaluation of the reliability of your imputation approach. This could take the form of just dropping varying proportions of missing data from your testing set and seeing how well your imputation approach performs. This would also give you a good basis for improving it in the future.
