### 2019-10-25 brittanyblouin/ANCRTAdjust#15 @seabbs
Great - thanks.

What are your thoughts about adding this to `data_clean`. In my experience users often don't read docs and it can be nice to give them very clear error messages when things go wrong.
