### 2019-10-25 brittanyblouin/ANCRTAdjust#14 @seabbs
Thanks - paper is much improved.

I still have a few issues with it however.

I would really like to see some discussion (and I may have missed this) of why standardised cleaning is needed over ad-hoc analyst by analyst cleaning (this may seem obvious but can be helpful for potential readers/users). For example, I may like to impute my missing data in a different way to the approach you have adopted (MICE), why should the users trust yours?

If it were me I would focus on the fact that you have standardised data cleaning in an open manner. Your approach may not, ultimately, be the optimal one but having this open source toolkit will allow a community of best practise to form. They (and you) can then easily make changes that are well documented and tested.
