### 2019-10-25 #1740 @seabbs
Thanks for these changes @m-maheu-giroux - I think the package is looking really good!

I've just gone through the review issues and added some more comments. I think the main work left to do is making sure unit test coverage is good enough, confirming package functionality, and making the need for the package very clear (nearly there on this).
