### 2019-10-25 brittanyblouin/ANCRTAdjust#21 @ellessenne
+1 for checking on R-Hub - you may get some weird errors/notes when testing on different platforms.
On a side note, I would actually suggest using `rhub::check_for_cran()` instead of the function from the `devtools` package (I think many functions from `devtools` are being relocated to other packages? Not sure if this is one of those).
