### 2019-10-25 brittanyblouin/ANCRTAdjust#18 @ellessenne
I would suggest including a simulated dataset that mimics characteristics of the data you are not allowed to share, and demonstrating the pipeline using that.
