### 2019-10-25 brittanyblouin/ANCRTAdjust#17 @ellessenne
I also see there are no documented examples when running R CMD check - you may want to add some to the documentation.
