### 2019-10-25 brittanyblouin/ANCRTAdjust#13 @ellessenne
I would suggest not using deprecated packages - they may suddenly stop working with new releases of R. If it is not too hard to move on from `plyr`, I would recommend doing so.

On a side note, `plyr` is imported twice in the DESCRIPTION file - which causes the following issue when running R CMD check:

```
❯ checking DESCRIPTION meta-information ... NOTE
  Package listed in more than one of Depends, Imports, Suggests, Enhances:
    ‘plyr’
  A package should be listed in only one of these fields.
```
