### 2019-11-05 #1740 @m-maheu-giroux
Hi @ellessenne and @seabbs,

I will do my best to address your suggestions shortly. However, my co-author is on maternity leave and I am traveling extensively over the next month. I anticipate some delays in incorporating your suggestions.

Thank you for your understanding,
Mathieu MG
