### 2020-10-25 #1740 @m-maheu-giroux
Hi,

Thanks for your patience.
The intent was still to refine the submission by addressing the few outstanding comments by the reviewers. However, my co-authors and I are not in a position to do so within the next few weeks. We unfortunately have to agree with your proposition to close (reject) the submission.

Again, I would like to thank @ellessenne  and @seabbs  for their time reviewing the package. We highly appreciated your inputs.

Best,
Mathieu MG
